move core of handleMessage logic into dataModel
